Updated deleteBranch for cleanUp

+++ routes/deleteBranch.js
@@ -5,6 +5,10 @@
 // Misc:
 const preRouteChecks = require('../utilities/preRouteChecks');
 const pushChecker = require('../utilities/pushChecker');
+const cleanUp = require('../utilities/cleanUp');
+const addToIPFS = require('../utilities/addToIPFS');
+const pushToBare = require('../utilities/pushToBare');
+
 
 const { exec } = require('child_process');
 
@@ -61,11 +65,55 @@ async function main(projName, username, timestamp, branchToUpdate, barerepopath,
         });
     } catch(err) {
         console.log(err);
-        //await rmWorkdir(workdirpath);
+        let checkExistBranch = await gitListBranches(workdirpath, branchName)
+        if (!checkExistBranch){
+            await gitBranchAdd(barerepopath, workdirpath, branchName);
+            await pushToBare(barerepopath, workdirpath, branchName);    
+        }
+        await addToIPFS(barerepopath);
+        await cleanUp(workdirpath, err.message);
         throw new Error(`(deleteBranch) main err ${err.name} :- ${err.message}`);
     }
 }
-
+async function gitBranchAdd(workdirpath, branchName) {
+    try {
+        await git.branch({
+            fs: fs,
+            dir: workdirpath,
+            ref: branchName,
+            checkout: true
+        })
+        return(true);
+    } catch(err) {
+        console.log(err); 
+        throw new Error(`git-branch err ${err.name} :- ${err.message}`);
+    }
+}
+async function gitListBranches(workdirpath, branchName) {
+    // I have to bring in remote branch names as well as local branch names (removing "HEAD" from this because isomorphic returns this for local)
+    try {
+        var branchlist = [];
+        let remoteBranches = await git.listBranches({
+            fs: fs,
+            dir: workdirpath,
+            remote: 'origin'
+        })
+        let localBranches = await git.listBranches({
+            fs:fs,
+            dir: workdirpath
+        })
+        branchlist = localBranches.concat(remoteBranches);
+        branchlist = branchlist.filter(branchname => branchname != "HEAD")
+        branchlist = branchlist.filter((v, i, a) => a.indexOf(v) === i); // Removing duplicates - credits - https://stackoverflow.com/a/14438954
+        if (branchlist.includes(branchName))
+            return true
+        else 
+            return false 
+    } catch (err) {
+        console.log(err);
+        throw new Error(`(gitListBranches) git-list-branch err ${err.name} :- ${err.message}`);
+    }
+}
 function removeBranchPath(branchPathToRemove) {
     return new Promise((resolve, reject) => {
         try {
@@ -84,22 +132,17 @@ function removeBranchPath(branchPathToRemove) {
 }
 async function gitDeleteBranch(workdirpath, branchName){
     try {
-        await git.deleteBranch({
-            fs: fs,
-            dir:  workdirpath,
-            ref: branchName,
-        })
-        return(true);
+        let checkExistBranch = await gitListBranches(workdirpath, branchName);
+        if (checkExistBranch){
+            await git.deleteBranch({
+                fs: fs,
+                dir:  workdirpath,
+                ref: branchName,
+            })
+        }
     }catch(err){
         if (err.name == "NotFoundError"){ 
-            /**
-             * 1. We don't have two branches available at the same time. 
-             *    Even to do so, its difficult because the route has to checkout this 
-             *    'existing' branch first to make it available.
-             * 2. So isomorphic-git will give this error which can be handled by allowing 
-             *    the route to later delete the branch at remote.
-             *  
-            */ 
+            await git.deleteRef({ fs: fs, dir: workdirpath, ref: `refs/origin/${branchName}` })
            return(true);
         } else {
             throw new Error(`deleteBranch git err ${err.name} :- ${err.message}`)
